add xml for connecting to cluster new tag

+++ src/main/scala/com/dfossouo/scala/hbase/HBaseWriter.scala
@@ -11,10 +11,11 @@ import org.apache.hadoop.hbase.util.Bytes
 object HBaseWriter {
 
   def createTableOrOverwrite(admin: Admin, table: HTableDescriptor): Unit = {
-/*    if (admin.tableExists(table.getTableName)) {
+
+    if (admin.isTableAvailable(table.getTableName)) {
       admin.disableTable(table.getTableName)
       admin.deleteTable(table.getTableName)
-    } */
+    }
     admin.createTable(table)
   }
